+++ src/js/ui/tab-maps.js
@@ -155,14 +155,19 @@ const exportSelectedMap = async () => {
 
 	const dir = ExportHelper.getExportPath(path.join('maps', selectedMapDir));
 
+	// The export helper provides the user with a link to the directory of the last exported
+	// item. Since we're using directory paths, we just append another segment here so that
+	// when the path is trimmed, users end up in the right place. Bit hack-y, but quicker.
+	const markPath = path.join('maps', selectedMapDir, selectedMapDir);
+
 	for (const index of exportTiles) {
 		const adt = new ADTExporter(selectedMapID, selectedMapDir, index);
 
 		try {
 			await adt.export(dir, exportQuality);
-			helper.mark(adt.tileID, true);
+			helper.mark(markPath, true);
 		} catch (e) {
-			helper.mark(adt.tileID, false, e.message);
+			helper.mark(markPath, false, e.message);
 		}
 	}