+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed issue that prevented some files from being available in remote mode.
 - Fixed issue in Blender plugin that slowed down imports significantly.
 - Fixed issue in Blender plugin where it would crash on models without materials.
+- The 'Open Export Directory' link after exporting has been replaced with 'View in Explorer' which opens the directory of the last exported item.
 
 0.1.8 (11-01-2020)
 - NPC variant textures can now be selected/exported for creature models.